+++ unity/PitayaExample/Assets/Pitaya/IPitayaSerializer.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Pitaya
+{
+    public interface IPitayaSerializer
+    {
+        byte[] Encode(object obj);
+        T Decode<T>(byte[] buffer);
+    }
+}