+++ unity/PitayaExample/Assets/Tests/Integration/JsonSSLRequestTest.cs
@@ -7,6 +7,7 @@ using Pitaya.SimpleJson;
 
 namespace Pitaya.Tests
 {
+    [TestFixture, Category("integration")]
     public class JsonSSLRequestTest
     {
 
@@ -25,9 +26,9 @@ namespace Pitaya.Tests
         [SetUp]
         public void Init()
         {
-            // PitayaBinding.AddPinnedPublicKeyFromCaFile("client-ssl.localhost.crt");
+            // StaticPitayaBinding.AddPinnedPublicKeyFromCaFile("client-ssl.localhost.crt");
             _mainThread = Thread.CurrentThread;
-            PitayaBinding.SkipKeyPinCheck(true);
+            StaticPitayaBinding.SkipKeyPinCheck(true);
 
             _client = new PitayaClient("myCA.pem");
             PitayaClient.SetLogLevel(PitayaLogLevel.Debug);