+++ CHANGELOG.md
@@ -5,9 +5,32 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for strongly typed JSON de/serialization using Newtonsoft.Json
+- Deprecated methods explicitly using JsonObject and IMessage
+- IPitayaClient interface
 - Better documentation
 - Unit testing (not only integration tests)
 - Better support for languages other than C#
+
+## [4.1.1] - Unreleased
+### Fixed
+- Fix crash in Unity 2019.3 when running in Android
+
+## [4.1.0] - 2020-06-03
+### Added
+- ClearAllCallbacks method on PitayaClient class
+
+### Fixed
+- Fix crash in Unity 2019.3 when running in Linux
+
+## [4.0.0] - 2020-04-04
+**This version breaks compatibility with Unity 2018.4 or older**
+### Fixed
+- Fix AndroidJavaClass errors for Pitaya.dll in Android
+
+### Added
+- Add UnityEngine.AndroidJNIModule dependency (this breaks compatibility with Unity versions older than 2019.3)
 
 ## [3.0.4] - 2020-02-04
 ### Fixed